Playing with the naming of the folders
The folders in which they need to edit are long.  We should correctly
point them to the exact location they need to edit, but we don't know
what those folders will be exactly.  Playing with splitting the folder
in two parts "where is the addon" versus "which file in the addon".
Seems to look best from the things I tried.